Move SPI for patch libraries to be an internal
Avoiding new SPI for patch libraries until we finalize
the configuration attribute name and placement.
