+++ dev/com.ibm.websphere.appserver.spi.classloading/bnd.bnd
@@ -11,7 +11,7 @@
 #     IBM Corporation - initial API and implementation
 #*******************************************************************************
 -include= ~../cnf/resources/bnd/bundle.props
-bVersion: 1.5
+bVersion: 1.4
 
 Bundle-Name: WebSphere Classloading SPI
 Bundle-Description: WebSphere Classloading SPI, version ${bVersion}
