+++ dev/com.ibm.websphere.appserver.spi.classloading/com.ibm.websphere.appserver.spi.classloading.pom
@@ -27,5 +27,5 @@
   <artifactId>com.ibm.websphere.appserver.spi.classloading</artifactId>
   <version>${bFullVersion}</version>
   <name>WebSphere Classloading SPI</name>
-  <description>WebSphere Classloading SPI, version 1.5</description>
+  <description>WebSphere Classloading SPI, version 1.4</description>
 </project>
