+++ dev/com.ibm.ws.classloading/src/com/ibm/ws/classloading/internal/ClassLoaderConfigurationExtended.java
@@ -1,20 +1,22 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2025 IBM Corporation and others.
+ * Copyright (c) 2025 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
  * 
  * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-/**
- * @version 1.5.0
- */
-@org.osgi.annotation.versioning.Version("1.5.0")
-@TraceOptions(traceGroup = "ClassLoadingService", messageBundle = "com.ibm.ws.classloading.internal.resources.ClassLoadingServiceMessages")
-package com.ibm.wsspi.classloading;
+package com.ibm.ws.classloading.internal;
 
-import com.ibm.websphere.ras.annotation.TraceOptions;
+import java.util.List;
+
+import com.ibm.wsspi.classloading.ClassLoaderConfiguration;
+
+/**
+ * This is an internal until we can agree on the attribute name for patch libraries
+ */
+public interface ClassLoaderConfigurationExtended {
+    ClassLoaderConfiguration setPatchLibraries(List<String> libs);
+    List<String> getPatchLibraries();
+}
